[chunks] Move tx_pool out of ShardsManager.
